refactor: :recycle: convert browser package to typescript
