--- packages/browser/src/notifier.d.ts
@@ -1,2 +1,0 @@
-export { default } from '../types/bugsnag'
-export * from '../types/bugsnag'
